add token expire unit test

+++ auth_demo/service_layer/services.py
@@ -23,6 +23,10 @@ class RoleNotExists(Exception):
     pass
 
 
+class RoleInUse(Exception):
+    pass
+
+
 class PasswordError(Exception):
     pass
 
@@ -59,6 +63,8 @@ def delete_role(name, role_repo):
     role = role_repo.get_by_name(name)
     if not role:
         raise RoleNotExists('role not exsits')
+    if len(role.users) > 0:
+        raise RoleInUse('role in use')
     role_repo.delete(role)
